Changed collision detection to work in the reverse direction
The collision detection was only transferring momentum when it was in the forward (positive) direction. To overcome this the code looks at the sings of the velocities. If one of them are negative, the particles need to be flipped on which takes on u2 and which takes on u1. If they aer both negative there just needs to be a sign flip for the quadratic formula to work as intended. If the discriminant is negative (and thus the solution is imaginary, this means that the particles are moving away from another and nothing needs to occur.